requirements.txt
discord.py[voice]
flask
flask-socketio
eventlet
yt-dlp
spotipy
cryptography
python-dotenv
